requirements/base.txt
requests~=2.24.0
PyYAML~=5.3.1
